[#3] 필터 컴포넌트 학과 선택 코드 컴포넌트로 분리
