+++ src/components/shared/Filter/Tab2.tsx
@@ -4,7 +4,7 @@ import { INTEREST } from 'src/schema';
 import { CheckBoxOrRadioNone, CommonTag, LabelContainer } from 'src/theme/CommonStyle';
 import { SubmitHandler } from 'react-hook-form';
 import { ButtonContainer } from './Tab1';
-import { FilterInput } from '../type';
+import { FilterInput, FilterProps } from '../type';
 
 const Container = styled(LabelContainer)`
   padding: 2rem 0;
@@ -22,7 +22,7 @@ const ExtendsButtonContainer = styled(ButtonContainer)`
   transform: translateX(-50%);
 `;
 
-const Tab2 = ({ register, handleSubmit }) => {
+const Tab2 = ({ register, handleSubmit, setOpen }: FilterProps) => {
   const [activeInterest, setActiveInterest] = useState([]);
 
   const onClickInterest = (e) => {
