+++ src/components/shared/type.ts
@@ -1,3 +1,6 @@
+import React, { SetStateAction } from 'react';
+import { UseFormHandleSubmit, UseFormRegister } from 'react-hook-form';
+
 export interface ActiveCheckBoxOrRadioColor {
   activeColor: boolean;
 }
@@ -8,4 +11,13 @@ export interface FilterInput {
   smoke?: string;
   military?: string;
   interest?: string[];
+}
+
+export interface FilterSetOpen {
+  setOpen: React.Dispatch<SetStateAction<boolean>>;
+}
+
+export interface FilterProps extends FilterSetOpen {
+  register?: UseFormRegister<FilterInput>;
+  handleSubmit: UseFormHandleSubmit<FilterInput>;
 }
